Add .py files and .ttf via upload
Temper.py is the minimalist implementation of Temperance. GoldenRatio.py includes a more informative interface to give a more game-like experience (keeping track of time and other data, and displaying that data for the user). SourceCodePro-Regular.ttf is the font used in Temperance. It was retrieved from Google's repository of fonts. 

To Follow: upload of executable files generated via cx_Freeze.